+++ MVC_Project.Domain/Services/ServiceBase.cs
@@ -28,7 +28,7 @@ namespace MVC_Project.Domain.Services
             _baseRepository.Create(entity);
         }
 
-        public void Delete(int id)
+        public void Delete(Int64 id)
         {
             _baseRepository.Delete(id);
         }
@@ -48,7 +48,7 @@ namespace MVC_Project.Domain.Services
             return _baseRepository.GetAll().ToList();
         }
 
-        public M GetById(int id)
+        public M GetById(Int64 id)
         {
             return _baseRepository.GetById(id);
         }
